Delete OOP boilerplate
So, the thing about OOP, and especially creating *multiple layers of inheritance* at once, is it really locks you down.
You have to think about all the interfaces and ownership at once before you can get a working system, and then it takes more work to change the architecture when you have so many individual points that rely on specific properties.
It might be fine in the end but how are you going to get there?

It would be better to start with one class of object, like GrypeTextItem, and get it working with all the behavior needed, and *then* split it up into a hierarchy of GrypeItem > GrypeWindingItem > GrypeTextItem in order to reuse behaviors for other item types.
But it might be better still not to assume any classes.