Position native input on top of path
Switch to a new strategy where the native input is positioned under the cursor at a dynamic position where the native input's text lines up with the visual text.

The goal is to enable more native behaviors, like double-click to select words, triple-click to select all, context menus, drag and drop of text, and on mobile, long-press to select.

That said, selection start/end handles will have to be reimplemented, and it's looking like it might be impossible to position the input at the start of a touch and get the native behaviors to apply to the input at the computed position.
But maybe the target is the only thing that is determined already,
and we can make the input large and use clip-path to shape it
to match the path, and then update the position and it'll use the
new position for native behavior? One can hope.

This is a work in progress, so the input is made visible for now,
and many behaviors are very broken.