Add `touch-action: none` to native input
hm, before:
- long press to start word selection and drag in same gesture: native input is not being repositioned while dragging, so selection is linear, but weirdly it seems limited to... the very top edge of the input? (mind you, the top is not where the text is, it's centered with `line-height`... though it could be assuming a standard line height, I suppose...)
- after selecting a word, drag where textbox is: can trigger scrolling, pull to refresh

after:
- long press to start word selection and drag in same gesture: native input is still not being repositioned while dragging, so selection is linear, but doesn't seem limited to the top of the input
- after selecting a word, drag where textbox is: repositions input... but stops after a bit??*

The native behavior for long press and drag in same gesture is that after selecting a word it allows changing the start or end of the selection, depending on which direction you drag it. (The side that is the anchor vs caret is chosen as soon as you start dragging; it doesn't change when you bring the selection end back to where you started.)

*I see native behavior if the text field is horizontal, and I
stay over the native input. It seems that when the input is
rotated, that's when it aborts the gesture.